Refactor attendance marking method and input handling
Handles the marking method : If student is present for the present or stipulated date mark 'P' if not mark 'A' else if a value or something else is inputed into system display 'invalid input, enter 'A'  for absent or 'P' for present. After the conditions are met/satisfied then print "attendance is marked successfully".